refactor : 요금관련 클래스들을 적절한 패키지 위치로 변경
