+++ src/main/java/wooteco/subway/docs/구현목록.md
@@ -9,7 +9,7 @@
 - [x] FareCalculator가 하고 있는 작업을 다른 객체들에게 분배시키기
   - [x] AgePolicy에서 제공하던 isTeenage와 같은 작업을 밖에서 호출하지 않고, 이곳에서 할인금액을 적용시키도록 수정
   - [x] 거리에따른 요금정책 적용을 DistancePolicy에서 수행하도록 수정
-- [ ] Policy 패키지 위치 변경
+- [x] Policy 패키지 위치 변경
 - [ ] 요금 계산 부분 단위 테스트 구현
 - [ ] 테스트에서 중복되어 사용되는 값들을 상위 클래스로 올리기
 - [ ] BeforeEach 위치 변경
